+++ src/namicorn.test.ts
@@ -58,7 +58,7 @@ describe('ZNS', () => {
       '0xaa91734f90795e80751c96e682a321bb3c1a4186',
     );
     expect(result.meta.owner).toEqual(
-      '0x267ca17e8b3bbf49c52a4c3b473cdebcbaf9025e',
+      'zil1ye72zl5t8wl5n3f2fsa5w0x7hja0jqj7mhct23',
     );
     expect(result.meta.type).toEqual('zns');
     expect(result.meta.ttl).toEqual(0);